headers/deps: ACPI: Optimize <linux/acpi.h> dependencies, remove <linux/irqdomain.h> inclusion
There's a lot of inclusions of <linux/irqdomain.h> through <linux/acpi.h>:

   2,190 In file included from ./include/linux/acpi.h:13

In fact this is over 65% of all irqdomain.h inclusions:

   3,369 include/linux/irqdomain.h

The dependency is spurious: the only reference to struct irq_domain_ops
can be expressed with a pre-declaration.

Fix up dangling dependencies of code that learned to rely on the
spurious <linux/irqdomain.h> inclusion.

Interestingly almost all ~90 of these dangling depenedencies were on
<linux/of_api.h>, while only one of them required <linux/irqdomain.h>
itself.

Signed-off-by: Ingo Molnar <mingo@kernel.org>