headers/deps: sched/wait: Remove <linux/swait_api.h> from <linux/swait.h>
Turn <linux/swait.h> into a data-types (and initializers) only header.

Add <linux/swait_api.h> in the places that are using the swait APIs.

          --------------------------------------------------------------------
          | Combined, preprocessed C code size of header, without line markers,
          | with comments stripped:
          -------------------------
  before: | #include <linux/swait.h>             | LOC:  8502 | headers: 220 |
   after: | #include <linux/swait.h>             | LOC:  8486 | headers: 219 |

Header complexity went down only in a minor fashion, because we still have
the full header dependencies in <linux/swait.h>. The following commits are
going to remove most of them.

Signed-off-by: Ingo Molnar <mingo@kernel.org>