headers/prep: Work around <linux/vmstat.h> header dependency bug temporarily
Currently <linux/vmstat.h> cannot be included before <linux/mm.h>:

	./include/linux/vmstat.h: In function ‘__mod_lruvec_page_state’:
	./include/linux/vmstat.h:519:24: error: implicit declaration of function ‘page_pgdat’; did you mean ‘page_private’? [-Werror=implicit-function-declaration]
	  519 |  __mod_node_page_state(page_pgdat(page), idx, val);
	      |                        ^~~~~~~~~~
	      |                        page_private

We are going to fix this bug in the fast-headers tree, but before that we
need to clean up various dependencies - which includes the addition of
explicit <linux/vmstat.h> dependencies.

To make that easier to script, to not have to watch the inclusion dependency
<linux/vmstat.h> has on <linux/mm.h>, include <linux/mm.h> temporarily,
to work around the bug.

This has to be done outside the usual header guard, because <linux/mm.h>
includes <linux/vmstat.h>.

This kludge will go away after all changes of the fast-headers tree is applied,
and has no negative effects on the current tree, as <linux/mm.h> had to be
included before <linux/vmstat.h> already.

Signed-off-by: Ingo Molnar <mingo@kernel.org>