headers/deps: Add initial new headers as identity mappings
As part of tree we are going to 'split' of a lot of header
files into 3 parts:

    <header.h>
    <header_types.h>
    <header_api.h>

To help this conversion we'll add in the over 20,000 new
dependencies to .c files in advance, to reduce the size and
maintenance complexity of the actual dependency optimization
patches.

Start out by introducing a select set of <header_api.h> and
<header_types.h> headers, which map back to the <header.h>
header.

These headers can from now on be used, without breaking the build.

Also add a few other new headers - all mapping back to their
originating header for the time being.

Signed-off-by: Ingo Molnar <mingo@kernel.org>