headers/deps: x86/fpu: Make task_struct::thread constant size
Turn thread.fpu into a pointer. Since most FPU code internals work by passing
around the fpu pointer, the code generation impact is small.

This allows us to remove the old kludge of task_struct being variable size,
which creates a number of problems, such as requiring thread_struct to be
the last member of the struct - not allowing it to be struct-randomized, etc.

The fpu_thread_struct_whitelist() quirk to hardened usercopy can be removed,
now that the FPU structure is not embedded in the task struct anymore.

Signed-off-by: Ingo Molnar <mingo@kernel.org>