headers/deps: syscalls: Split <linux/syscalls_api.h> out of <linux/syscalls.h>
In a typical distro build over 170 files include the <linux/syscalls.h> header:

       175 In file included from ./include/linux/syscalls.h:79,

And it's a huge header:

          --------------------------------------------------------------------
          | Combined, preprocessed C code size of header, without line markers,
          | with comments stripped:
          -------------------------
          | #include <linux/syscalls.h>          | LOC: 34816 | headers: 545 |

But the overwhelming majority of usage sites only needs this header for
the sycall defines, not for the syscall prototypes or helper inlines
offered by the header.

Split the latter out into a separate API header: <linux/syscalls_api.h>.

In this patch the new header is still included in the main header file,
to not break existing code - that inclusion will be removed in the next
patch.

Signed-off-by: Ingo Molnar <mingo@kernel.org>