headers/deps: hashtable_types: Use ilog2_const()
Make HASH_BITS() use ilog2_const().

Since it uses ARRAY_SIZE(), which is not supposed to be used on dynamic arrays,
this shouldn't be a limitation in practice.

If some code uses it that way, a compiler error or warning should alert us,
due to HASH_BITS() returning -1. I hope. :-/

Signed-off-by: Ingo Molnar <mingo@kernel.org>